# backend/openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "backend | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/parties/{partyId}/memory/graph": {
      "get": {
        "tags": [
          "Memory"
        ],
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryGraphDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryGraphDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryGraphDto"
                }
              }
            }
          }
        }
      }
    },
    "/parties/{partyId}/memory/participants/{personaId}/stances": {
      "post": {
        "tags": [
          "Memory"
        ],
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "personaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendStanceRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendStanceRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AppendStanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AppendStanceResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppendStanceResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppendStanceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Memory"
        ],
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "personaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StanceRecord"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StanceRecord"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StanceRecord"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/parties/{partyId}/memory/participants/{personaId}/stances/{stanceId}/retract": {
      "post": {
        "tags": [
          "Memory"
        ],
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "personaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "stanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AppendStanceResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppendStanceResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppendStanceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/parties/{partyId}/memory/participants/{personaId}/consolidate": {
      "post": {
        "tags": [
          "Memory"
        ],
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "personaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConsolidationRunResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsolidationRunResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsolidationRunResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/parties/{partyId}/memory/consolidate": {
      "post": {
        "tags": [
          "Memory"
        ],
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsolidationRunResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsolidationRunResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsolidationRunResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppendStanceRequest": {
        "required": [
          "valence",
          "reasoning",
          "targetKind",
          "targetPersonaId",
          "targetConceptName"
        ],
        "type": "object",
        "properties": {
          "valence": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "reasoning": {
            "type": [
              "null",
              "string"
            ]
          },
          "targetKind": {
            "$ref": "#/components/schemas/StanceTargetKind"
          },
          "targetPersonaId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "targetConceptName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "AppendStanceResponse": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ConsolidationRunResponse": {
        "required": [
          "runId",
          "personaId",
          "recollectionsWalked",
          "stancesAppended",
          "watermark",
          "skipped"
        ],
        "type": "object",
        "properties": {
          "runId": {
            "type": "string",
            "format": "uuid"
          },
          "personaId": {
            "type": "string",
            "format": "uuid"
          },
          "recollectionsWalked": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "stancesAppended": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "watermark": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "skipped": {
            "type": "boolean"
          }
        }
      },
...
      "MemoryGraphDto": {
        "required": [
          "nodes",
          "links"
        ],
        "type": "object",
        "properties": {
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemoryGraphNode"
            }
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemoryGraphLink"
            }
          }
        }
      },
      "MemoryGraphLink": {
        "required": [
          "source",
          "target",
          "kind"
        ],
        "type": "object",
        "properties": {
          "source": {
            "type": "string"
          },
          "target": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "snippet": {
            "type": [
              "null",
              "string"
            ]
          },
          "ts": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "MemoryGraphNode": {
        "required": [
          "id",
          "kind"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "display": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": [
              "null",
              "string"
            ]
          },
          "roomId": {
            "type": [
              "null",
              "string"
            ]
          },
          "anchorMessageId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "StanceOrigin": {
        "enum": [
          "Curator",
          "Consolidation",
          "Retract"
        ],
        "default": "Curator"
      },
      "StanceRecord": {
        "required": [
          "id",
          "valence",
          "reasoning",
          "ts",
          "targetKind",
          "targetPersonaId",
          "targetConceptName",
          "targetConceptDisplay",
          "isCurrent"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "valence": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "reasoning": {
            "type": "string"
          },
          "ts": {
            "type": "string",
            "format": "date-time"
          },
          "targetKind": {
            "$ref": "#/components/schemas/StanceTargetKind"
          },
          "targetPersonaId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "targetConceptName": {
            "type": [
              "null",
              "string"
            ]
          },
          "targetConceptDisplay": {
            "type": [
              "null",
              "string"
            ]
          },
          "isCurrent": {
            "type": "boolean"
          },
          "origin": {
            "$ref": "#/components/schemas/StanceOrigin"
          },
          "runId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "retractsId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
      "StanceTargetKind": {
        "enum": [
          "Participant",
          "Concept",
          "Self"
        ]
      },
...
    }
  },
...
}